+++ CHANGELOG.md
@@ -3,10 +3,14 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 
+
 ## 1.1 - 20-06-2023
 
 ### Fixed
 - fixed let variables not being shown
+
+### Changed
+- improved scope naming
 
 
 ## 1.0 - 17-06-2023